+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](http://semver.org/) and is followi
 
 ## Unreleased
 
+### :zap: Added
+
+- [#528](https://github.com/FantasticFiasco/aws-signature-version-4/issues/528) Add support for `HttpClient.PatchAsync` (contribution by [@mungojam](https://github.com/mungojam))
+
 ### :skull: Removed
 
 - [BREAKING CHANGE] Support for .NET Framework 4.5 due to [deprecation as of January 12, 2016](https://docs.microsoft.com/en-us/lifecycle/products/microsoft-net-framework)